+++ backend/file-ucloud-service/src/main/resources/db/migration/V15__LastScan.sql
@@ -1,3 +1,0 @@
-create table if not exists file_ucloud.storage_scan (
-    last_system_scan timestamp not null
-)